function/trend_fetcher/requirements.txt
boto3
xdk
python-ulid
